rename AutoTranslateLiveActionConfig to AutoTranslateCaConfig since it applies to workflow and actions
