+++ aem/core/pom.xml
@@ -45,7 +45,7 @@
                             <bnd><![CDATA[
 Export-Package: com.composum.ai.aem.core.impl.autotranslate
 Import-Package: javax.annotation;version=0.0.0,*
-Sling-ContextAware-Configuration-Classes: com.composum.ai.aem.core.impl.autotranslate.rollout.AutoTranslateLiveActionConfig
+Sling-ContextAware-Configuration-Classes: com.composum.ai.aem.core.impl.autotranslate.AutoTranslateCaConfig
 ]]></bnd>
                         </configuration>
                     </execution>
