--- src/site/markdown/autogenerated/parts/slingca.AutoTranslateLiveActionConfig.md.part
@@ -1,9 +1,0 @@
-/* AIGenVersion(234a7c96, slingcaconfigurations.prompt-395da17f, AutoTranslateLiveActionConfig.java-26cef814) */
-
-## Composum AI Automatic Translation POC Rollout Configuration (aem-core)
-
-Configures rollout details for automatic translation. This is a demo and not yet fully functional.
-
-| id                      | label                    | type   | default value | description                                       |
-|-------------------------|--------------------------|--------|---------------|---------------------------------------------------|
-| additionalInstructions  | Additional Instructions  | String |               | Additional instructions for the automatic translation. |
